Make requirements.txt optional and check VERNEED instead of symbols

+++ pyfreezer/freeze.sh
@@ -1,6 +1,9 @@
-#!/bin/bash -e
+#!/bin/bash
+set -euo pipefail
 
-./pysrc/python -m pip install -r $(dirname $1)/requirements.txt
+if [[ -f "$(dirname $1)/requirements.txt" ]]; then
+    ./pysrc/python -m pip install -r $(dirname $1)/requirements.txt
+fi
 ./pysrc/python ./pysrc/Tools/freeze/freeze.py -p $(realpath pysrc) -e $(realpath ext) -o build "$@"
 cp overrides.c build/
 export LDFLAGS="overrides.c -Wl,-Bstatic"
@@ -8,9 +11,9 @@ cd build && make LIBS="-Wl,-Bdynamic -lrt -lutil -lpthread -ldl" && cd -
 cp build/$(basename $1 .py) $(dirname $1)
 
 # Checks
-GLIBC_VER=$(readelf --dyn-syms --wide build/$(basename $1 .py) | grep GLIBC_ | grep -v WEAK | cut -d@ -f2 | cut -d_ -f2 | cut -d' ' -f1 | sort -V | tail -1)
+GLIBC_VER=$(readelf -V build/$(basename $1 .py) | grep -Po 'Name: GLIBC_\d+.\d+.\d*' | grep -Po '\d+\.\d+\.?\d*' | sort -V | tail -1)
 echo "Earliest supported glibc: $GLIBC_VER"
-readelf --syms build/$(basename $1 .py) | grep GLIBC_${GLIBC_VER}
+readelf --dyn-syms --wide build/$(basename $1 .py) | grep GLIBC_${GLIBC_VER}
 
 echo "ldd:"
 ldd build/$(basename $1 .py)
